added future classes. Simplified GCS add drone to observe ack behavior
