+++ Drone/gcs.py
@@ -1,0 +1,1 @@
+## future GCS class
